Improved Documentation and folder structure
