+++ Python3_Version/data/blue_stain_temp_and_growth_rate.csv
